+++ webapp/src/components/Controls/Controls.tsx
@@ -3,9 +3,11 @@ import {
   Box,
   Button,
   CircularProgress,
+  FormControlLabel,
   IconButton,
   InputAdornment,
   OutlinedInput,
+  Switch,
   Typography,
   useTheme,
 } from "@mui/material";
@@ -13,6 +15,7 @@ import {
   QueryFilterState,
   QueryPaginationState,
   QueryPipelineState,
+  QueryPostprocessingState,
 } from "types/models";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import { motion } from "framer-motion";
@@ -42,6 +45,7 @@ type Props = {
   filters: QueryFilterState;
   pagination: QueryPaginationState;
   pipeline: QueryPipelineState;
+  postprocessing: QueryPostprocessingState;
   searchString: string;
 };
 
@@ -49,6 +53,7 @@ const Controls: React.FC<Props> = ({
   filters,
   pagination,
   pipeline,
+  postprocessing,
   searchString,
 }) => {
   const theme = useTheme();
@@ -69,6 +74,7 @@ const Controls: React.FC<Props> = ({
           datasetSplitName,
           ...filters,
           ...pipeline,
+          ...postprocessing,
         })
       : getUtteranceCountPerFilterEndpoint.useQuery({
           jobId,
@@ -101,7 +107,11 @@ const Controls: React.FC<Props> = ({
 
   const handleClearFilters = () => {
     history.push(
-      `${baseUrl}${constructSearchString({ ...pagination, ...pipeline })}`
+      `${baseUrl}${constructSearchString({
+        ...pagination,
+        ...pipeline,
+        ...postprocessing,
+      })}`
     );
   };
 
@@ -111,6 +121,7 @@ const Controls: React.FC<Props> = ({
         ...filters,
         ...pagination,
         ...pipeline,
+        ...postprocessing,
       })}`
     );
 
@@ -128,6 +139,16 @@ const Controls: React.FC<Props> = ({
 
   const handleDatasetSplitChange = (name: DatasetSplitName) =>
     history.push(`/${jobId}/dataset_splits/${name}/${mainView}${searchString}`);
+
+  const handlePostprocessingChange = (enable: boolean) =>
+    history.push(
+      `${baseUrl}${constructSearchString({
+        ...filters,
+        ...pagination,
+        ...pipeline,
+        withoutPostprocessing: enable || undefined,
+      })}`
+    );
 
   const transition = { type: "tween" };
 
@@ -205,6 +226,17 @@ const Controls: React.FC<Props> = ({
             <DatasetSplitToggler
               value={datasetSplitName}
               onChange={handleDatasetSplitChange}
+            />
+          </Box>
+          <Box margin={1}>
+            <FormControlLabel
+              control={
+                <Switch
+                  checked={postprocessing.withoutPostprocessing ?? false}
+                  onChange={(_, checked) => handlePostprocessingChange(checked)}
+                />
+              }
+              label="Without PostProcessing"
             />
           </Box>
           <Box