+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - New Smart Tags `pipeline_disagreement` and `incorrect_for_all_pipelines` as a first step for pipeline comparison.
 - Links on top words to filter utterances that contain it.
 - `/confusion_matrix` now supports `normalized:bool` to get counts versus normalized confusion matrix.
+- added "Without Postprocessing" feature to enable/disable post processing feature on the data
 
 ### Changed